requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3 
